all: fix array size data type on 64 bit
The array struct aligned size to 64 bit on x86_64 which caused bad things
to happen.

We use size_t to make sure the size element is correct regardless arch.

Solution found by Timo.

+++ src/apk_defines.h
@@ -87,7 +87,7 @@ void *apk_array_resize(void *array, size_t new_size, size_t elem_size);
 
 #define APK_ARRAY(array_type_name, elem_type_name)			\
 	struct array_type_name {					\
-		int num;						\
+		size_t num;						\
 		elem_type_name item[];					\
 	};								\
 	static inline void						\
